refactor: exported SavedDecks into a separate component
