+++ components/navigation/Decks.tsx
@@ -1,0 +1,7 @@
+import SavedDecks from "../decks/SavedDecks";
+
+const Decks: React.FC = () => {
+  return <SavedDecks />;
+};
+
+export default Decks;
